requirements.txt
# ===== Core Deep Learning Frameworks =====
# PyTorch with CUDA support (for GTX 1650)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# ===== Computer Vision =====
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
ultralytics>=8.0.0  # YOLOv8
pillow>=10.0.0

# ===== Screen Capture =====
mss>=9.0.0
pyautogui>=0.9.54
pyscreeze>=0.1.29

# ===== Reinforcement Learning =====
stable-baselines3>=2.0.0
gymnasium>=0.29.0  # OpenAI Gym 的繼任者
sb3-contrib>=2.0.0

# ===== Android Automation =====
pure-python-adb>=0.3.0
# Alternatively: pip install adb-shell

# ===== Data Processing =====
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ===== Visualization =====
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
plotly>=5.14.0

# ===== Configuration & Utilities =====
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
loguru>=0.7.0

# ===== OCR (Optional) =====
pytesseract>=0.3.10
easyocr>=1.7.0

# ===== Testing =====
pytest>=7.4.0
pytest-cov>=4.1.0

# ===== Code Quality =====
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
